added support for LocaleString

+++ builder.py
@@ -11,7 +11,7 @@ single_part = None
 if len(sys.argv) > 1:
 	single_part = sys.argv[1]
 
-if config.overwrite_db and os.path.isfile(dotabase_db) and (single_part is None):
+if __name__ == "__main__" and config.overwrite_db and os.path.isfile(dotabase_db) and (single_part is None):
 	print("overwriting db")
 	os.remove(dotabase_db)
 session = dotabase_session()
